edit controller

+++ identityprovider/src/main/java/nz/ac/canterbury/seng302/identityprovider/model/GroupRepository.java
@@ -12,7 +12,7 @@ import java.util.List;
  */
 @Repository
 public interface GroupRepository extends CrudRepository<Groups, Long> {
-    Groups findByGroupId(long id);
+    Groups findByGroupId(int id);
 
     List<Groups> findAllByGroupShortName(String shortName);
     List<Groups> findAllByGroupLongName(String longName);
